Perform victory condition. Not sure why the score isn't updated yet...

+++ src/com/voyagegames/monkeymatch/helpers/Token.java
@@ -9,9 +9,9 @@ public class Token {
 	public final int type;
 	public final Vector2 initialPosition;
 	public final float weight;
-	public final float value;
+	public final int value;
 	
-	public Token(final Actor actor, final int type, final Vector2 initialPosition, final float weight, final float value) {
+	public Token(final Actor actor, final int type, final Vector2 initialPosition, final float weight, final int value) {
 		this.actor = actor;
 		this.type = type;
 		this.initialPosition = initialPosition;
